+++ backend/package.json
@@ -21,6 +21,7 @@
     "@types/cors": "^2.8.18",
     "@types/express": "^5.0.1",
     "@types/jest": "^29.5.14",
+    "@types/sinon": "^17.0.4",
     "axios": "^1.9.0",
     "cors": "^2.8.5",
     "crypto": "^1.0.1",
@@ -28,6 +29,7 @@
     "jest": "^29.7.0",
     "nyc": "^17.1.0",
     "pg-promise": "^11.13.0",
+    "sinon": "^20.0.0",
     "ts-jest": "^29.3.2",
     "ts-node": "^10.9.2",
     "typescript": "^5.8.3"
