--- backend/test/integration/signup.test.ts
@@ -1,75 +1,0 @@
-import Signup from "../../src/Signup";
-import { AccountDAOMemory } from "../../src/AccountDAO";
-import GetAccount from "../../src/GetAccount";
-
-let signup: Signup;
-let getAccount: GetAccount;
-
-beforeEach(() => {
-  const accountDAO = new AccountDAOMemory();
-  signup = new Signup(accountDAO);
-  getAccount = new GetAccount(accountDAO);
-})
-
-test("Deve criar uma conta válida", async () => {
-  const inputSignup = {
-    name: "John Doe",
-    email: "john.doe@gmail.com",
-    document: "97456321558",
-    password: "asdQWE123"
-  };
-
-  const outputSignup = await signup.execute(inputSignup);
-
-  expect(outputSignup.accountId).toBeDefined();
-
-  const outputGetAccount = await getAccount.execute(outputSignup.accountId);
-
-  expect(outputGetAccount.name).toBe(inputSignup.name);
-  expect(outputGetAccount.email).toBe(inputSignup.email);
-  expect(outputGetAccount.document).toBe(inputSignup.document);
-});
-
-test("Não deve criar uma conta com nome inválido", async () => {
-  const inputSignup = {
-    name: "John",
-    email: "john.doe@gmail.com",
-    document: "97456321558",
-    password: "asdQWE123"
-  };
-
-  await expect(() => signup.execute(inputSignup)).rejects.toThrow("Invalid name");
-});
-
-test("Não deve criar uma conta com email inválido", async () => {
-  const inputSignup = {
-    name: "John Doe",
-    email: "john.doe",
-    document: "97456321558",
-    password: "asdQWE123"
-  };
-
-  await expect(() => signup.execute(inputSignup)).rejects.toThrow("Invalid email");
-});
-
-test("Não deve criar uma conta com CPF inválido", async () => {
-  const inputSignup = {
-    name: "John Doe",
-    email: "john.doe@gmail.com",
-    document: "974563215",
-    password: "asdQWE123"
-  };
-
-  await expect(() => signup.execute(inputSignup)).rejects.toThrow("Invalid document");
-});
-
-test("Não deve criar uma conta com senha inválida", async () => {
-  const inputSignup = {
-    name: "John Doe",
-    email: "john.doe@gmail.com",
-    document: "97456321558",
-    password: "asdQWE"
-  };
-
-  await expect(() => signup.execute(inputSignup)).rejects.toThrow("Invalid password");
-});
